refactor: deduplicate shared utilities and fix cookie store robustness
- Extract `getSesskey()` into `services/sesskey.ts` (was duplicated in scraper.ts and dashboard.ts)
- Extract `isLoginHtml()` into `utils/moodle-url.ts` (was copy-pasted across 4 service files)
- Fix cookie store: validate Date before assigning expires (malformed Set-Cookie headers created Invalid Date objects that were never cleaned up)

No behavioral changes — all imports updated, TypeScript compiles clean.
